+++ rust/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### NEXT
 
-- Worker: Remove `io_uring` support (PR #XXXX).
+- Worker: Remove `io_uring` support (PR #1832).
 
 ### 0.22.6