+++ rust/CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### 0.22.5
 
-- Unify all consumer classes into a single one ([PR #1731](https://github.com/versatica/mediasoup/pull/1731)).
+- Worker: Unify all `Consumer` classes into a single one ([PR #1731](https://github.com/versatica/mediasoup/pull/1731)).
 - SCTP: Validate CRC32c checksum in received packets ([PR #1828](https://github.com/versatica/mediasoup/pull/1828)).
 - SCTP: Authenticate State Cookie in plain and pipe transports ([PR #1829](https://github.com/versatica/mediasoup/pull/1829)).